backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.2.2
Pillow==10.2.0
django-filter==23.5
ipython==8.18.1
weasyprint==61.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1